# frontend/lib/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SupportNova",
  "description": "\n**SupportNova** turns a customer complaint into a routed, escalated,\npolicy-grounded resolution \u2014 where a deterministic Python rule engine\nindependently verifies and, where required, overrides the AI before anything\nreaches the customer.\n\nTwo pipelines run over the same inputs:\n\n* **Pipeline 1 \u2014 GenAI Complaint Intelligence** (`genai_pipeline/`)\n  classifies, extracts and drafts, returning strictly-schemaed JSON.\n* **Pipeline 2 \u2014 Python Ground-Truth Validation** (`python_validation/`)\n  reaches its *own* conclusion from the Complaint Resolution Rule Matrix,\n  using no language model at all.  It can be run from the command line with\n  no API key configured.\n\nTheir outputs are reconciled field by field; Python holds veto on department,\nurgency, priority, escalation, policy validity and prohibited actions.\n",
  "contact": {
   "name": "SupportNova Team"
  },
  "license": {
   "name": "MIT"
  },
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/auth/register": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Create a customer account and sign in",
    "description": "Open sign-up, rate-limited per IP like login.\n\nThe account created is always a **customer**: the request cannot ask for\na role, and the service ignores anything that tries. Staff accounts come\nfrom an administrator, so no one can promote themselves by posting JSON.",
    "operationId": "register_api_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/login": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Log in and receive an access + refresh token pair",
    "description": "Rate-limited to `RATE_LIMIT_LOGIN` per IP.  Failures are audited but the\nerror message never distinguishes an unknown account from a bad password.",
    "operationId": "login_api_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/refresh": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Rotate an expiring session",
    "operationId": "refresh_api_auth_refresh_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RefreshRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/auth/logout": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Revoke a refresh token",
    "operationId": "logout_api_auth_logout_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RefreshRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MessageResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/auth/me": {
   "get": {
    "tags": [
     "Authentication"
    ],
    "summary": "The signed-in user",
    "operationId": "me_api_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/auth/change-password": {
   "post": {
    "tags": [
     "Authentication"
    ],
    "summary": "Change your own password",
    "operationId": "change_password_api_auth_change_password_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PasswordChangeRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MessageResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
  "/api/audit": {
   "get": {
    "tags": [
     "Audit"
    ],
    "summary": "The organisation-wide audit trail, newest first",
    "description": "Every recorded act, filterable by what was touched, what was done and who\ndid it.\n\n``before`` and ``after`` are returned as stored: two snapshots, not a diff.\nSRS Step 59 wants the original recommendation and the reviewer's decision\nboth on the record, and a diff would keep only what changed.",
    "operationId": "list_audit_api_audit_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Size"
      }
     },
     {
      "name": "entity_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 64
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Type"
      }
     },
     {
      "name": "action",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 48
        },
        {
         "type": "null"
        }
       ],
       "title": "Action"
      }
     },
     {
      "name": "actor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 320
        },
        {
         "type": "null"
        }
       ],
       "description": "Actor email.",
       "title": "Actor"
      },
      "description": "Actor email."
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 64
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_AuditEntryOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/audit/actions": {
   "get": {
    "tags": [
     "Audit"
    ],
    "summary": "Every action that has been recorded, with a count",
    "description": "What the trail contains, for a filter control and for the dashboard's\n\"what changed\" figure. Read from the rows, so an action that has never\nhappened is absent rather than listed at zero.",
    "operationId": "audit_actions_api_audit_actions_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AuditActionCountOut"
         },
         "type": "array",
         "title": "Response Audit Actions Api Audit Actions Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/audit/{entity_type}/{entity_id}": {
   "get": {
    "tags": [
     "Audit"
    ],
    "summary": "One entity's full trail, oldest first",
    "description": "How one thing got to its current state.\n\nOldest first, because a trail is read forwards. Capped, because an entity\nwith more than five hundred recorded acts is a bug report rather than a\npage.",
    "operationId": "entity_trail_api_audit__entity_type___entity_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "entity_type",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity Type"
      }
     },
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AuditEntryOut"
         },
         "title": "Response Entity Trail Api Audit  Entity Type   Entity Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "AuditActionCountOut": {
    "properties": {
     "action": {
      "type": "string",
      "title": "Action"
     },
     "count": {
      "type": "integer",
      "title": "Count"
     }
    },
    "type": "object",
    "required": [
     "action",
     "count"
    ],
    "title": "AuditActionCountOut",
    "description": "An action name and how often it has been recorded."
   },
   "AuditEntryOut": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "At"
     },
     "actor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor"
     },
     "actor_role": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actor Role"
     },
     "entity_type": {
      "type": "string",
      "title": "Entity Type"
     },
     "entity_id": {
      "type": "string",
      "title": "Entity Id"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "before": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Before"
     },
     "after": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "After"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "request_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Request Id"
     }
    },
    "type": "object",
    "required": [
     "id",
     "entity_type",
     "entity_id",
     "action"
    ],
    "title": "AuditEntryOut",
    "description": "One recorded act.\n\n``before`` and ``after`` are the stored snapshots, not a diff: SRS Step 59\nwants the original recommendation and the reviewer's decision both on the\nrecord, and a diff would keep only what changed between them."
   },
...
   "DepartmentRef": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "code",
     "name"
    ],
    "title": "DepartmentRef"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LoginRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "LoginRequest"
   },
   "MessageResponse": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "message"
    ],
    "title": "MessageResponse"
   },
...
   "Page_AuditEntryOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/AuditEntryOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page",
      "default": 1
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size",
      "default": 25
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[AuditEntryOut]"
   },
...
   "PasswordChangeRequest": {
    "properties": {
     "current_password": {
      "type": "string",
      "title": "Current Password"
     },
     "new_password": {
      "type": "string",
      "maxLength": 256,
      "minLength": 8,
      "title": "New Password"
     }
    },
    "type": "object",
    "required": [
     "current_password",
     "new_password"
    ],
    "title": "PasswordChangeRequest"
   },
...
   "RefreshRequest": {
    "properties": {
     "refresh_token": {
      "type": "string",
      "title": "Refresh Token"
     }
    },
    "type": "object",
    "required": [
     "refresh_token"
    ],
    "title": "RefreshRequest"
   },
   "RegisterRequest": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "maxLength": 255,
      "minLength": 2,
      "title": "Full Name"
     },
     "password": {
      "type": "string",
      "maxLength": 256,
      "minLength": 12,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "email",
     "full_name",
     "password"
    ],
    "title": "RegisterRequest",
    "description": "Self-service sign-up.\n\nThere is deliberately no ``role`` field. A role decides what a person may\nchange in a complaint register, so it is not something the person applying\nfor the account gets to choose \u2014 the server always assigns CUSTOMER, and\nstaff roles are provisioned by an administrator."
   },
...
   "TokenResponse": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "refresh_token": {
      "type": "string",
      "title": "Refresh Token"
     },
     "token_type": {
      "type": "string",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In"
     },
     "user": {
      "$ref": "#/components/schemas/UserOut"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "refresh_token",
     "expires_in",
     "user"
    ],
    "title": "TokenResponse"
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "department": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/DepartmentRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "is_active",
     "created_at"
    ],
    "title": "UserOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "JWT access token",
    "scheme": "bearer"
   }
  }
 }
}